Slider Values Added
There are now values shown next to their associated sliders.

Important note: There's an issue with the distance that the values hang from their associated labels (this is more noticeable if you zoom out and in). I did a workaround that kind of fixes it, but not really (I just added some spaces in the HTML file).

Basically, the values do not share the exact same position horizontally.